Remove initDownloadService from ArtifactoryServicesManager
The interface ArtifactoryServicesManager contains the private method
initDownloadService.  This means that gomock for example cannot be used
to generate a set of mocks for this interface which can then be used in
unit-testing.  In this change we simply remove initDownloadService from
ArtifactoryServicesManager, which means that it is now only defined and
used by ArtifactoryServicesManagerImp.