Add ability to ignore hash check failures.
Filtered resources manipulate the contents of artifacts during download.
The SHA1 supplied by the server does not reflect those changes, but
instead only hashes the raw unfiltered template. In the current
implementation, the client always checks that the hash matches the
expected hash and returns an error if they do not match.

This change adds a flag to indicate whether the hash should be
validated defaulting to false).