Encode properties in predictable order
When testing, it is useful to be able to have some predictability in the
way that properties are encoded to a string. Since the properties struct
is currently backed by a map, the order that it is rendered to a string
is not predictable.

This change sorts the keys before rendering to a string. The items in
the arrays (the values in the map) are considered to be ordered and
hence are not sorted.

Feature Request: https://github.com/jfrog/jfrog-client-go/issues/549